requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.18
python-jose[cryptography]==3.3.0
google-auth==2.29.0
httpx==0.27.0
deepface==0.0.92
tf-keras==2.16.0
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.7.1
google-generativeai==0.7.2
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
psycopg2-binary>=2.9.9
